requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.2.0
anthropic>=0.25.0

# PDF parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Data validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0

# Optional: for better PDF parsing with formula support
# pdf2image>=1.16.0
# pymupdf>=1.23.0
